Fixed a bug in the Utils class
Utils class had a method that returns a local IP address of the
connected device.  It was returning null because my conditional
statements were wrong.  I  fixed the issue by removing an if statement
that checks if the instance of the app is already running.  There was no
need for that check, and currently it works.  Tested it by connecting
two devices physically.