Re-Programmed the forget wi-fi direct group
Before the forget group network feature needed user input to actually
delete the group.  The app only redirected them to the wifi settings and
the user had to find their way around and figure out how to delete the
group.  With this way now, I designed it so that by a click of a button
the groups will be deleted and requires no other user action.