Added new screen for options and re-arranged the layout
Re-arranged the main home layout of the buttons so the most important
button such as 'Connect' is shown first in the grid to capture the
users' attention.  I also created a new layout for the Options feature
where the user can perform certain tasks (to be done later).  Currently
the layout is only complete for the options feature.