Added a feature so that any file type can be sent from Client to Server
Before this feature the client could only send an image file to the
server but with this new feature now, the client can send various types
of files to the server.  Once the server device accepts the file, it is
up to the server device to choose how to open that file i.e. to open a
.PDF the server must have a .PDF reader app.