Added a feature where all files except images can be sent and recieved
Before I was only send a image from device A to device B for testing
purposes.  As the testing phase of that componenet was complete, I added
a feature so that all files can be sent from device A to device B.  The
only type of files that cannot is image files.  App is not intended for
images.